REFACTOR: change model module in schedule persistence layer
