--- schedule/persistence/src/main/kotlin/vp/togedo/document/PersonalScheduleDocument.kt
@@ -1,212 +1,0 @@
-package vp.togedo.document
-
-import com.fasterxml.jackson.annotation.JsonProperty
-import org.bson.types.ObjectId
-import org.springframework.data.annotation.Id
-import org.springframework.data.mongodb.core.index.Indexed
-import org.springframework.data.mongodb.core.mapping.Document
-import reactor.core.publisher.Mono
-import vp.togedo.enums.ScheduleEnum
-import vp.togedo.util.exception.schedule.ConflictScheduleException
-import vp.togedo.util.exception.schedule.EndTimeBeforeStartTimeException
-import vp.togedo.util.exception.schedule.InvalidTimeException
-import vp.togedo.util.exception.schedule.ScheduleNotFoundException
-
-@Document(collection = "personal_schedule")
-data class PersonalScheduleDocument(
-    @Id
-    @JsonProperty("id")
-    val id: ObjectId = ObjectId.get(),
-
-    @Indexed(unique = true)
-    @JsonProperty("userId")
-    val userId: ObjectId,
-
-    @JsonProperty("fixedSchedules")
-    val fixedSchedules: MutableList<Schedule> = mutableListOf(),
-
-    @JsonProperty("flexibleSchedules")
-    val flexibleSchedules: MutableList<Schedule> = mutableListOf()
-){
-
-    fun deleteFixedScheduleById(scheduleId: ObjectId): Mono<PersonalScheduleDocument> {
-        return Mono.fromCallable {
-            findIndexFixedScheduleById(scheduleId)
-        }.map{
-            fixedSchedules.removeAt(it)
-        }.map{
-            this
-        }
-    }
-
-    fun deleteFlexibleScheduleById(scheduleId: ObjectId): Mono<PersonalScheduleDocument> {
-        return Mono.fromCallable {
-            findIndexFlexibleScheduleById(scheduleId)
-        }.map{
-            flexibleSchedules.removeAt(it)
-        }.map{
-            this
-        }
-    }
-
-    fun modifyFixedSchedule(schedule: Schedule): Mono<PersonalScheduleDocument>{
-        return Mono.fromCallable{findIndexFixedScheduleById(schedule.id)}
-            .flatMap { scheduleIndex ->
-                deleteFixedScheduleByIndex(scheduleIndex)
-                    .flatMap { originalSchedule ->
-                        addFixedSchedule(schedule)
-                            .doOnError {
-                                fixedSchedules.add(scheduleIndex, originalSchedule)
-                            }
-                    }
-            }
-    }
-
-    fun modifyFlexibleSchedule(schedule: Schedule): Mono<PersonalScheduleDocument>{
-        return Mono.fromCallable{findIndexFlexibleScheduleById(schedule.id)}
-            .flatMap { scheduleIndex ->
-                deleteFlexibleScheduleByIndex(scheduleIndex)
-                    .flatMap { originalSchedule ->
-                        addFlexibleSchedule(schedule)
-                            .doOnError {
-                                flexibleSchedules.add(scheduleIndex, originalSchedule)
-                            }
-                    }
-            }
-    }
-
-    private fun findIndexFixedScheduleById(scheduleId: ObjectId): Int{
-        val index = fixedSchedules.indexOfFirst {it.id == scheduleId}
-        if (index < 0)
-            throw ScheduleNotFoundException("해당 스케줄이 존재하지 않습니다.")
-        return index
-    }
-
-    private fun findIndexFlexibleScheduleById(scheduleId: ObjectId): Int{
-        val index = flexibleSchedules.indexOfFirst {it.id == scheduleId}
-        if (index < 0)
-            throw ScheduleNotFoundException("해당 스케줄이 존재하지 않습니다.")
-        return index
-    }
-
-    private fun deleteFixedScheduleByIndex(index: Int): Mono<Schedule>{
-        return Mono.fromCallable {
-            fixedSchedules.removeAt(index)
-        }
-    }
-
-    private fun deleteFlexibleScheduleByIndex(index: Int): Mono<Schedule>{
-        return Mono.fromCallable {
-            flexibleSchedules.removeAt(index)
-        }
-    }
-
-    fun addFixedSchedule(schedule: Schedule): Mono<PersonalScheduleDocument> {
-        return Mono.fromCallable {
-            this.checkFixedScheduleValidTime(schedule)
-
-            val insertedIndex = this.getInsertedIndex(schedule, ScheduleEnum.FIXED_PERSONAL_SCHEDULE)
-
-            this.fixedSchedules.add(insertedIndex, schedule)
-
-            this
-        }
-    }
-
-    fun addFlexibleSchedule(schedule: Schedule): Mono<PersonalScheduleDocument> {
-        return Mono.fromCallable {
-            this.checkFlexibleScheduleValidTime(schedule)
-
-            val insertedIndex = this.getInsertedIndex(schedule, ScheduleEnum.FLEXIBLE_PERSONAL_SCHEDULE)
-
-            this.flexibleSchedules.add(insertedIndex, schedule)
-
-            this
-        }
-    }
-
-    fun getInsertedIndex(schedule: Schedule, scheduleEnum: ScheduleEnum): Int{
-        val schedules: MutableList<Schedule> = when(scheduleEnum) {
-            ScheduleEnum.FIXED_PERSONAL_SCHEDULE -> fixedSchedules
-            ScheduleEnum.FLEXIBLE_PERSONAL_SCHEDULE -> flexibleSchedules
-        }
-
-        if(schedules.size == 0)
-            return 0
-
-        val index = schedules.binarySearch(0){
-            scheduleElement -> scheduleElement.startTime.compareTo(schedule.startTime)
-        }
-
-        if (index >= 0)
-            throw ConflictScheduleException("해당 시작시간에 시작하는 스케줄이 있습니다.")
-
-        val insertedIndex = (index + 1) * -1
-
-        //앞의 종료 시간과, 해당 스케줄의 시작 시간을 비교
-        if(insertedIndex > 0)
-            if(schedules[insertedIndex - 1].endTime >= schedule.startTime)
-                throw ConflictScheduleException("전 스케줄이 종료되지 않았습니다.")
-
-        //뒤의 시작 시간과, 해당 스케줄의 종료 시간을 비교
-        if(insertedIndex < schedules.size)
-            if(schedules[insertedIndex].startTime <= schedule.endTime)
-                throw ConflictScheduleException("뒤 스케줄의 시작시간과 충돌합니다.")
-
-        return insertedIndex
-    }
-
-    fun checkFixedScheduleValidTime(schedule: Schedule): Boolean{
-        return this.checkScheduleValidTime(schedule, 1_00_00, 7_23_59)
-    }
-
-    fun checkFlexibleScheduleValidTime(schedule: Schedule): Boolean{
-        //년(2)_월(2)_일(2)_시(2)_분(2)
-        return this.checkScheduleValidTime(schedule, 10_01_01_00_00, 99_12_31_23_59)
-    }
-
-    fun checkScheduleValidTime(schedule: Schedule, startTimeRange: Long, endTimeRange: Long): Boolean {
-        this.isStartTimeBefore(schedule)
-
-        try{
-            isValidTime(schedule.startTime, startTimeRange, endTimeRange)
-        }catch (e: InvalidTimeException){
-            throw InvalidTimeException("시작 시간이 ${e.message}")
-        }
-
-        try{
-            isValidTime(schedule.endTime, startTimeRange, endTimeRange)
-        }catch (e: InvalidTimeException){
-            throw InvalidTimeException("종료 시간이 ${e.message}")
-        }
-
-        return true
-    }
-
-    fun isValidTime(time: Long, startTimeRange: Long, endTimeRange: Long): Boolean{
-        if(time !in startTimeRange .. endTimeRange)
-            throw InvalidTimeException("시간 범위 밖입니다.")
-
-        val hour = (time % 10000) / 100
-        if (hour !in 0 .. 23)
-            throw InvalidTimeException("hour 범위 밖입니다.")
-        val minute = time % 100
-        if (minute !in 0..59)
-            throw InvalidTimeException("minute 범위 밖입니다.")
-        return true
-    }
-
-    fun isStartTimeBefore(schedule: Schedule): Boolean{
-        if (schedule.startTime > schedule.endTime)
-            throw EndTimeBeforeStartTimeException("종료 시간이 시작시간보다 앞입니다.")
-        return true
-    }
-}
-
-data class Schedule(
-    val id: ObjectId,
-    val startTime: Long,
-    val endTime: Long,
-    val title: String,
-    val color: String
-)
