--- schedule/persistence/src/main/kotlin/vp/togedo/enums/ScheduleEnum.kt
@@ -1,6 +1,0 @@
-package vp.togedo.enums
-
-enum class ScheduleEnum {
-
-    FIXED_PERSONAL_SCHEDULE, FLEXIBLE_PERSONAL_SCHEDULE
-}
