+++ schedule/persistence/src/main/kotlin/vp/togedo/repository/GroupRepository.kt
@@ -2,7 +2,7 @@ package vp.togedo.repository
 
 import org.bson.types.ObjectId
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository
-import vp.togedo.document.GroupDocument
+import vp.togedo.model.documents.group.GroupDocument
 
 interface GroupRepository: ReactiveMongoRepository<GroupDocument, ObjectId> {
 }
