+++ schedule/persistence/src/main/kotlin/vp/togedo/repository/mongo/JoinedGroupMongoRepository.kt
@@ -2,7 +2,7 @@ package vp.togedo.repository.mongo
 
 import org.bson.types.ObjectId
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository
-import vp.togedo.document.JoinedGroupDocument
+import vp.togedo.model.documents.joinedGroup.JoinedGroupDocument
 
 interface JoinedGroupMongoRepository: ReactiveMongoRepository<JoinedGroupDocument, ObjectId> {
 }
