+++ schedule/persistence/src/main/kotlin/vp/togedo/repository/mongo/PersonalScheduleMongoRepository.kt
@@ -2,10 +2,7 @@ package vp.togedo.repository.mongo
 
 import org.bson.types.ObjectId
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository
-import reactor.core.publisher.Mono
-import vp.togedo.document.PersonalScheduleDocument
+import vp.togedo.model.documents.personalSchedule.PersonalScheduleDocument
 
 interface PersonalScheduleMongoRepository: ReactiveMongoRepository<PersonalScheduleDocument, ObjectId> {
-
-    fun findByUserId(userId: ObjectId): Mono<PersonalScheduleDocument>
 }
